requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
jinja2==3.1.6
pydantic==2.11.7
python-multipart==0.0.20
requests==2.32.4
python-pptx==1.0.2
